requirements.txt
python-dotenv
Pillow